change header&hash data type. was: real,double is: log,longint

+++ utils/c/include/shunt_primitives.h
@@ -117,7 +117,7 @@ typedef s_vpi_vecval svLogicVecVal;
   Returns: 
   hash value
 */
-double shunt_prim_hash(const char *str);
+long shunt_prim_hash(const char *str);
 
 /*
   Function: shunt_prim_error
